Implement observer pattern in ClientSession.
The pattern allows other classes (its dependents) to listen to updates
in the ClientSession. The update event is triggered by a new selection
of the active ontology that may change the current active project and
the active version ontology.

If the active ontology has no link to the remote project then the active
project and the active version ontology are null. The caller code must
be aware of this null value.