Client refactor - part 1
The client interface specifies over 55 methods, clearly too large for more than one implementation.
Most of these are used by the server admin gui, some aren't used at all and many are called in one place,
making for longer path lengths that aren't necessary.

This commit separates out a new Config class, used primarily by the admin gui. The client hangs on to a reference
to it for purposes of saving it to the server. The client now focuses primarily on creating/deleting projects and
committing changes to the ontology, basically interactions with the server. As editors of the metaproject are now
required to connect using a different port, the assumption will be that if the admin tab is enabled the logged in
user has full privileges to edit, thereby removing the need for a policy mediator interface. This will be refactored
out in subsequent commits.