Make the handing over the ontology from server to Protege explicit.
The construction of an ontology coming from the server involves creating
an ontology baseline and updating it with the latest changes. This
process of creation is using a mock ontology manager. To be able for
Protege to recognize this ontology we need to "hand over" the ownership
from the mock ontology manager to Protge's ontology manager.